Logging: send logs to STDOUT (fixes #4)
This ensures that `delayed_jobs` will output to STDOUT which is more
suitable for use in a container. This is done in the `Dockerfile` and
modifies an upstream configuration file.

It also ensures that Rails logs got to STDOUT with a change to the
deployed version of the production environment configuration file.